rm redundant
